// src/interfaces/turnos/Turno.interface.ts
import { Types, Document } from 'mongoose';
import { Especialidad } from './TurnoEspecialidad.enum.ts';
import { EstadoTurno } from './TurnoEstado.enum.ts';

export interface ITurno extends Document {
    id?: Types.ObjectId;
    paciente: Types.ObjectId;
    especialidad: Especialidad;
    fechaTurno: Date;
    estado?: EstadoTurno;
    observaciones?: string;
    activo: boolean;
}